tests for make_dipole_forward
