+++ mne/simulation/tests/test_source.py
@@ -1,22 +1,14 @@
 import os.path as op
 
 import numpy as np
-from numpy.testing import (assert_array_almost_equal, assert_array_equal,
-                           assert_allclose)
-from nose.tools import assert_true, assert_raises
+from numpy.testing import assert_array_almost_equal, assert_array_equal
+from nose.tools import assert_true
 
 from mne.datasets import testing
-from mne import (read_label, read_forward_solution, pick_types_forward,
-                 pick_info, read_dipole, read_evokeds, read_cov, fit_dipole,
-                 pick_types, make_sphere_model)
+from mne import read_label, read_forward_solution, pick_types_forward
 from mne.label import Label
-from mne.dipole import Dipole
-from mne.proj import make_eeg_average_ref_proj
-from mne.simulation.evoked import simulate_evoked
-from mne.simulation.source import (simulate_stc, simulate_sparse_stc,
-                                   make_forward_dipole)
-from mne.source_estimate import VolSourceEstimate
-from mne.utils import run_tests_if_main, slow_test
+from mne.simulation.source import simulate_stc, simulate_sparse_stc
+from mne.utils import run_tests_if_main
 
 
 data_path = testing.data_path(download=False)
@@ -213,95 +205,5 @@ def test_simulate_sparse_stc_single_hemi():
 
     assert_array_equal(stc_1.lh_vertno, stc_2.lh_vertno)
     assert_array_equal(stc_1.rh_vertno, stc_2.rh_vertno)
-
-
-@slow_test
-@testing.requires_testing_data
-def test_make_forward_dipole():
-    """Test forward-projecting dipoles"""
-    rng = np.random.RandomState(0)
-
-    evoked = read_evokeds(fname_evo)[0]
-    info = evoked.info
-    cov = read_cov(fname_cov)
-    dip_c = read_dipole(fname_dip)
-
-    # Make new Dipole object with n_test_dipoles, uneven sampling in time
-    n_test_dipoles = 5
-    dipsel = np.sort(rng.permutation(range(len(dip_c)))[:n_test_dipoles])
-    dip_test = Dipole(times=dip_c.times[dipsel],
-                      pos=dip_c.pos[dipsel],
-                      amplitude=dip_c.amplitude[dipsel],
-                      ori=dip_c.ori[dipsel],
-                      gof=dip_c.gof[dipsel])
-
-    sphere = make_sphere_model(head_radius=0.1)
-    stc, fwd = make_forward_dipole(dip_test, sphere, evoked.info,
-                                   trans=fname_trans)
-    assert_true(isinstance(stc, list))
-
-    times, pos, amplitude, ori, gof = [], [], [], [], []
-    snr = 20
-    for s in stc:
-        evo_test = simulate_evoked(fwd, s, evoked.info, cov,
-                                   snr=snr, random_state=rng)
-        evo_test.add_proj(make_eeg_average_ref_proj(evo_test.info))
-        dfit, resid = fit_dipole(evo_test, cov, sphere, None)
-        times += list(dfit.times)
-        pos += list(dfit.pos)
-        amplitude += list(dfit.amplitude)
-        ori += list(dfit.ori)
-        gof += list(dfit.gof)
-
-    dip_fit = Dipole(np.array(times), np.array(pos), np.array(amplitude),
-                     np.array(ori), np.array(gof))
-
-    # check that we did at least as well
-    diff = dip_test.pos - dip_fit.pos
-    corr = np.corrcoef(dip_test.pos.ravel(), dip_fit.pos.ravel())[0, 1]
-    dist = np.sqrt(np.mean(np.sum(diff * diff, axis=1)))
-    gc_dist = 180 / np.pi * np.mean(np.arccos(
-            np.sum(dip_test.ori * dip_fit.ori, axis=1)))
-    amp_err = np.sqrt(np.mean((dip_test.amplitude - dip_fit.amplitude) ** 2))
-
-    # Make sure each coordinate is within 2 mm from reference
-    # NB tolerance should be set relative to snr of simulated evoked!
-    assert_allclose(dip_fit.pos, dip_test.pos, rtol=0, atol=2e-3,
-                    err_msg='position mismatch')
-    assert_true(dist < 2e-3, 'dist: %s' % dist)  # within 2 mm
-    assert_true(corr > 1 - 2e-3, 'corr: %s' % corr)
-    assert_true(gc_dist < 2, 'gc_dist: %s' % gc_dist)  # less than 2 degrees
-    assert_true(amp_err < 2e-3, 'amp_err: %s' % amp_err)  # within 2 nAm
-
-    # Make sure rejection works with bem: one dipole at z=1m
-    # NB _make_forward.py:_prepare_for_forward will raise a RuntimeError
-    # if no points are left after min_dist exclusions
-    dip_outside = Dipole(times=np.array([0., 0.001]),
-                         pos=np.array([[0., 0., 1.0], [0., 0., 0.040]]),
-                         amplitude=np.array([100e-9, 100e-9]),
-                         ori=np.array([[1., 0., 0.], [1., 0., 0.]]), gof=1)
-    assert_raises(ValueError, make_forward_dipole, dip_outside, fname_bem,
-                  evoked.info, fname_trans)
-
-    # Now make an evenly sampled set of dipoles, some simultaneous,
-    # should return a VolSourceEstimate regardless
-    times = np.array([0., 0., 0., 0.001, 0.001, 0.002])
-    pos = np.random.rand(6, 3)*0.020 + np.array([0., 0., 0.040])[np.newaxis, :]
-    amplitude = np.random.rand(6,)*100e-9
-    ori = np.eye(6, 3) + np.eye(6, 3, -3)
-    gof = np.arange(len(times))/len(times)
-
-    dip_even_samp = Dipole(times, pos, amplitude, ori, gof)
-
-    # restrict sensors to EEG using pre-picking
-    picks = pick_types(info, meg=False, eeg=True)
-    eeg_info = pick_info(evoked.info, picks)
-    stc, fwd = make_forward_dipole(dip_even_samp, fname_bem, eeg_info,
-                                   trans=fname_trans)
-
-    assert_true(isinstance, VolSourceEstimate)
-    assert_allclose(stc.times, np.arange(0., 0.003, 0.001))
-    assert_true(fwd['info']['ch_names'][0] == 'EEG 001')
-    assert_true(fwd['info']['ch_names'][-1] == 'EEG 060')
 
 run_tests_if_main()
